Put sketch in an appropriately named folder
The Arduino IDE requires that the sketch folder name matches the sketch file name. When you open a sketch that is in a folder with a mismatched name the Arduino IDE prompts the user and then moves the sketch file to an appropriately named folder. Unfortunately the Arduino IDE does not also move the other sketch files along with it. That causes the sketch to no longer compile as it requires all the other .ino files that were located in the src folder.

Closes https://github.com/Espressobar/arduino-grind-timer/issues/2